backend/requirements.txt
fastapi
uvicorn
google-generativeai
faiss-cpu
sentence-transformers
python-dotenv
pydantic
numpy
